requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
apex=0.1=pypi_0
appdirs=1.4.4=py_0
attrs=20.3.0=pyhd3eb1b0_0
black=19.10b0=py_0
blas=1.0=mkl
boto=2.49.0=py36_0
boto3=1.16.48=pyhd3eb1b0_0
botocore=1.19.48=pyhd3eb1b0_0
bpemb=0.3.2=pypi_0
brotlipy=0.7.0=py36h27cfd23_1003
ca-certificates=2020.12.8=h06a4308_0
cachetools=4.2.0=pyhd3eb1b0_0
certifi=2020.12.5=py36h06a4308_0
cffi=1.14.4=py36h261ae71_0
chardet=4.0.0=py36h06a4308_1003
click=7.1.2=py_0
configargparse=1.2.3=pypi_0
cryptography=3.3.1=py36h3c74f83_0
cudatoolkit=10.2.89=hfd86e86_1
cycler=0.10.0=py36_0
cython=0.29.21=py36he6710b0_0
dataclasses=0.8=pypi_0
dbus=1.13.18=hb2f20db_0
dill=0.3.2=pypi_0
docutils=0.15.2=py36_0
expat=2.2.10=he6710b0_2
filelock=3.0.12=pypi_0
fontconfig=2.13.0=h9420a91_0
freetype=2.10.4=h5ab3b9f_0
future=0.18.2=pypi_0
gensim=3.8.0=py36h962f231_0
glib=2.66.1=h92f7085_0
google-api-core=1.22.2=py36_0
google-auth=1.21.3=py_0
google-cloud-core=1.5.0=pyhd3eb1b0_0
google-cloud-storage=1.35.0=pyhd3eb1b0_0
google-crc32c=1.1.0=py36h27cfd23_1
google-resumable-media=1.2.0=pyhd3eb1b0_1
googleapis-common-protos=1.52.0=py36_0
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py36hd6299e0_1
h5py-cache=1.0=pypi_0
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=2.10=py_0
intel-openmp=2020.2=254
jmespath=0.10.0=py_0
joblib=1.0.0=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
kiwisolver=1.3.0=py36h2531618_0
lcms2=2.11=h396b838_0
ld_impl_linux-64=2.33.1=h53a641e_7
libcrc32c=1.1.1=he6710b0_2
libedit=3.1.20191231=h14c3975_1
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.13.0.1=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.2=heb0550a_3
matplotlib=3.3.1=0
matplotlib-base=3.3.1=py36h817c723_0
mkl=2020.2=256
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.2.0=py36h23d657b_0
mkl_random=1.1.1=py36h0573a6f_0
multidict=5.1.0=py36h27cfd23_2
mypy_extensions=0.4.3=py36_0
ncurses=6.2=he6710b0_1
ninja=1.10.2=py36hff7bd54_0
nlp=0.4.0=pypi_0
nltk=3.5=py_0
numpy=1.19.2=py36h54aff64_0
numpy-base=1.19.2=py36hfa32c7d_0
olefile=0.46=py36_0
openssl=1.1.1i=h27cfd23_0
packaging=20.8=pypi_0
pandas=1.1.3=py36he6710b0_0
pathspec=0.7.0=py_0
pcre=8.44=he6710b0_0
pillow=8.1.0=py36he98fc37_0
pillow-simd=7.0.0.post3=pypi_0
pip=20.3.3=py36h06a4308_0
portalocker=2.0.0=pypi_0
progressbar2=3.53.1=pypi_0
protobuf=3.13.0.1=py36he6710b0_1
pyarrow=1.0.1=pypi_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=py_0
pyqt=5.9.2=py36h05f1152_2
pysocks=1.7.1=py36h06a4308_0
python=3.6.12=hcff3b4d_2
python-dateutil=2.8.1=py_0
python-utils=2.4.0=pypi_0
pytorch=1.6.0=py3.6_cuda10.2.89_cudnn7.6.5_0
pytz=2020.5=pyhd3eb1b0_0
qt=5.9.7=h5867ecd_1
readline=8.0=h7b6447c_0
regex=2020.11.13=py36h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
rsa=4.6=py_0
s3transfer=0.3.3=pyhd3eb1b0_3
sacrebleu=1.4.14=pypi_0
sacremoses=0.0.43=pypi_0
scikit-learn=0.23.2=py36h0573a6f_0
scipy=1.5.2=py36h0b6359f_0
seaborn=0.11.0=py_0
sentencepiece=0.1.91=pypi_0
setuptools=51.0.0=py36h06a4308_2
sip=4.19.8=py36hf484d3e_0
six=1.15.0=py36h06a4308_0
smart_open=4.1.0=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tokenizers=0.9.4=pypi_0
toml=0.10.1=py_0
torchtext=0.7.0=pypi_0
torchvision=0.7.0=py36_cu102
tornado=6.1=py36h27cfd23_0
tqdm=4.50.2=py_0
transformers=4.1.0=pypi_0
typed-ast=1.4.1=py36h7b6447c_0
typing_extensions=3.7.4.3=py_0
urllib3=1.25.11=py_0
wheel=0.36.2=pyhd3eb1b0_0
xxhash=2.0.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.5.1=py36h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h9ceee32_0
